working rerender charts
