+++ src/game/MainGameScreen.js
@@ -2,9 +2,7 @@ import React, { useRef, useState, useEffect} from 'react';
 import ReactTimer from "@xendora/react-timer";
 import {socket} from '../LoginScreen';
 import PlayerStats from '../game/PlayerStats.js';
-import Charts from '../game/Charts';
-import styled from 'styled-components';
-import { Bar } from 'react-chartjs-2';
+import styled, { css } from 'styled-components';
 
 const prompt = "One study examining 30 subjects, of varying different styles and expertise, has found minimal difference in typing speed between touch typists and self-taught hybrid typists. According to the study, 'The number of fingers does not determine typing speed... People using self-taught typing strategies were found to be as fast as trained typists... instead of the number of fingers, there are other factors that predict typing speed... fast typists... keep their hands fixed on one position, instead of moving them over the keyboard, and more consistently use the same finger to type a certain letter.' To quote doctoral candidate Anna Feit: 'We were surprised to observe that people who took a typing course, performed at similar average speed and accuracy, as those that taught typing to themselves and only used 6 fingers on average' (Wikipedia)";
 
@@ -85,7 +83,7 @@ function MainGameScreen({playerName, room}) {
         <GirdItem>
           {promptJSX()}
           {gameStateJSX()}
-          <Input type="text" name="name" ref={textboxRef} onChange={onTextChanged} />
+          <Input type="text" disabled={playerFinished} name="name" ref={textboxRef} onChange={onTextChanged} />
         </GirdItem>
         <GirdItem>
           <p>WPM: {wpm}</p>
@@ -108,7 +106,7 @@ const GridContainer = styled.div`
 `;
 
 const GirdItem = styled.div`
-  padding: 50px;
+  padding: 40px;
 `;
 
 const Input = styled.textarea`
@@ -118,4 +116,7 @@ const Input = styled.textarea`
  &:focus{
   background-color: lightblue;
  }
+ ${({ disabled }) => disabled && css`
+    display: none;
+  `}
 `;
